+++ server/src/Communication/CommunicationManager.h
@@ -14,7 +14,7 @@ public:
     CommunicationManager(char *ipAddress, unsigned int port);
 
     Packet listen();
-    Packet putTimestamp(Packet packet, float timestamp);
+    Packet putTimestamp(Packet packet, int64_t timestamp);
 
     void handleConnection(sockaddr_in clientAddress);
 };
